adjust size type in itemcontainer to accommodate larger inventories

+++ src/servers/sapphire_zone/Inventory/ItemContainer.h
@@ -8,34 +8,34 @@
 namespace Sapphire
 {
 
-  using ItemMap = std::map< uint8_t, ItemPtr >;
+  using ItemMap = std::map< uint16_t, ItemPtr >;
 
   class ItemContainer
   {
 
   public:
-    ItemContainer( uint16_t storageId, uint8_t maxSize, const std::string& tableName, bool isMultiStorage,
+    ItemContainer( uint16_t storageId, uint16_t maxSize, const std::string& tableName, bool isMultiStorage,
                    bool isPersistentStorage = true );
 
     ~ItemContainer();
 
     uint16_t getId() const;
 
-    uint8_t getEntryCount() const;
+    uint16_t getEntryCount() const;
 
-    void removeItem( uint8_t slotId );
+    void removeItem( uint16_t slotId );
 
     ItemMap& getItemMap();
 
     const ItemMap& getItemMap() const;
 
-    ItemPtr getItem( uint8_t slotId );
+    ItemPtr getItem( uint16_t slotId );
 
-    void setItem( uint8_t slotId, ItemPtr item );
+    void setItem( uint16_t slotId, ItemPtr item );
 
-    int8_t getFreeSlot();
+    int16_t getFreeSlot();
 
-    uint8_t getMaxSize() const;
+    uint16_t getMaxSize() const;
 
     std::string getTableName() const;
 
@@ -45,7 +45,7 @@ namespace Sapphire
 
   private:
     uint16_t m_id;
-    uint8_t m_size;
+    uint16_t m_size;
     std::string m_tableName;
     bool m_bMultiStorage;
     bool m_isPersistentStorage;
